pyproject.toml
[project]
name = "protocol-lab"
version = "0.1.0"
description = "A hands-on lab for various application layer protocols."
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standard]",
    "fastapi",
    "sqlalchemy[asyncio]",
    "pydantic",
    "strawberry-graphql",
    "grpcio",
    "grpcio-tools",
    "websockets",
    "aiortc",
    "python-multipart", # Required by FastAPI for form data, good to include
    "httpx", # For client-side HTTP requests in benchmarks/clients
    "aiosqlite>=0.22.1",
    "rich>=14.3.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "streamlit>=1.54.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "psutil>=7.2.2",
    "faker>=40.1.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {where = ["src"]}
